organize gql queries
