+++ src/ActorLogic.ts
@@ -2,9 +2,9 @@ import { useActor } from "@/store/actor";
 import { useQuery } from "@vue/apollo-composable";
 import { storeToRefs } from "pinia";
 import { Ref, watch } from "vue";
-import { GetEmoteSet, GetEmoteSetMin } from "./assets/gql/emote-set/emote-set";
-import { GetCurrentUser } from "./assets/gql/users/self";
-import { GetUser } from "./assets/gql/users/user";
+import { GetEmoteSet, GetEmoteSetMin } from "@/apollo/query/emote-set.query";
+import { GetCurrentUser } from "./apollo/query/user-self.query";
+import { GetUser } from "./apollo/query/user.query";
 import { useObjectSubscription } from "./composable/object-sub";
 import { useModal } from "./store/modal";
 import { useMutationStore } from "./store/mutation";
