+++ src/App.vue
@@ -78,8 +78,8 @@ import { storeToRefs } from "pinia";
 import { provideApolloClient, useQuery } from "@vue/apollo-composable";
 import { SEASONAL_THEME_START, useStore } from "@store/main";
 import { tooltip } from "@/composable/tooltip";
-import { AppState, GetAppState } from "@gql/users/self";
-import { apolloClient } from "@/apollo";
+import { AppState, GetAppState } from "@/apollo/query/user-self.query";
+import { apolloClient } from "@/apollo/apollo";
 import { useI18n } from "vue-i18n";
 import { useContextMenu } from "./composable/context-menu";
 import { options } from "@/i18n";
