+++ src/apollo/mutation/message.mutation.ts
@@ -1,5 +1,15 @@
 import gql from "graphql-tag";
 
+export const ReadMessages = gql`
+	mutation ReadMessages($ids: [ObjectID!]!, $read: Boolean!) {
+		readMessages(message_ids: $ids, read: $read)
+	}
+`;
+
+export interface ReadMessages {
+	readMessages: number;
+}
+
 export const SendInboxMessage = gql`
 	mutation SendInboxMesage(
 		$recipients: [ObjectID!]!
