+++ src/apollo/mutation/report.mutation.ts
@@ -1,6 +1,20 @@
 import { Report } from "@/structures/Report";
 import gql from "graphql-tag";
 
+export const CreateReport = gql`
+	mutation CreateReport($data: CreateReportInput!) {
+		createReport(data: $data) {
+			id
+			subject
+			body
+		}
+	}
+`;
+
+export interface CreateReport {
+	createReport: Report;
+}
+
 export const EditReport = gql`
 	mutation EditReport($id: ObjectID!, $data: EditReportInput!) {
 		editReport(report_id: $id, data: $data) {
