+++ src/apollo/query/cosmetics.query.ts
