--- src/assets/gql/mutation/CreateReport.ts
@@ -1,16 +1,0 @@
-import { Report } from "@/structures/Report";
-import gql from "graphql-tag";
-
-export const CreateReport = gql`
-	mutation CreateReport($data: CreateReportInput!) {
-		createReport(data: $data) {
-			id
-			subject
-			body
-		}
-	}
-`;
-
-export interface CreateReport {
-	createReport: Report;
-}
