--- src/assets/gql/mutation/DeleteRole.ts
@@ -1,11 +1,0 @@
-import gql from "graphql-tag";
-
-export const DeleteRole = gql`
-	mutation DeleteRole($role_id: ObjectID!) {
-		deleteRole(role_id: $role_id)
-	}
-`;
-
-export interface DeleteRole {
-	deleteRole: string;
-}
