+++ src/views/emote/EmoteStats.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script setup lang="ts">
-import { GetEmote, GetEmoteStatistics } from "@/assets/gql/emotes/emote";
+import { GetEmote, GetEmoteStatistics } from "@/apollo/query/emote.query";
 import { useQuery } from "@vue/apollo-composable";
 import { computed } from "vue";
 import { useI18n } from "vue-i18n";
