+++ src/views/store/SubStatus.vue
@@ -114,9 +114,9 @@ import { useI18n } from "vue-i18n";
 import { useEgVault } from "./egvault";
 import { badgeDefs, getNextBadge, BadgeDef, getBadgeByID } from "@/components/utility/BadgeDefs";
 import { useQuery } from "@vue/apollo-composable";
-import { GetUserCosmetics } from "@/assets/gql/users/self";
+import { GetUserCosmetics } from "@/apollo/query/user-self.query";
 import { useActor } from "@/store/actor";
-import { GetUser } from "@/assets/gql/users/user";
+import { GetUser } from "@/apollo/query/user.query";
 import { Badge, Paint } from "@/structures/Cosmetic";
 import { useModal } from "@/store/modal";
 import differenceInDays from "date-fns/fp/differenceInDays";
